modify comments, remove unuse variable, set messages for all type of error.
solve rebase related problem
