+++ Trappist/src/Promact.Trappist.Core/Controllers/BasicSetupController.cs
@@ -43,7 +43,6 @@ namespace Promact.Trappist.Core.Controllers
         [HttpPost("mailsettings")]
         public async Task<IActionResult> ValidateEmailSettings([FromBody] EmailSettings model)
         {
-            return Ok(true);
             if (ModelState.IsValid)
                 return Ok(await _basicSetup.ValidateEmailSetting(model));
             return BadRequest();
