+++ Trappist/src/Promact.Trappist.Repository/Promact.Trappist.Repository.csproj
@@ -12,7 +12,6 @@
 
   <ItemGroup>
     <PackageReference Include="AutoMapper" Version="6.0.1" />
-    <PackageReference Include="MailKit" Version="1.12.0" />
     <PackageReference Include="System.Data.SqlClient" Version="4.3.0" />
   </ItemGroup>
