+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/delete-category-dialog.html
@@ -4,7 +4,7 @@
         <button type="button" class="pull-right" md-dialog-close><em class="material-icons">close</em></button>
     </div>
     <md-dialog-content>
-        <h4 class="mb20">Are you sure you want to delete this section?</h4>
+        <h4 class="mb20">Are you sure you want to delete {{category.categoryName}} Category?</h4>
     </md-dialog-content>
     <md-dialog-actions>
         <button type="button" class="btn" md-dialog-close>No</button>
