+++ Trappist/src/Promact.Trappist.Web/wwwroot/app/questions/questions-dashboard/delete-question-dialog.html
@@ -8,6 +8,6 @@
   </md-dialog-content>
   <md-dialog-actions>
     <button type="button" class="btn" md-dialog-close>No</button>
-    <button type="submit" class="btn btn-primary" >Yes</button>
+    <button type="submit" class="btn btn-primary">Yes</button>
   </md-dialog-actions>
 </div>
